Changed seriesNumber from int to float in recognition of series with books like 3.5. Adjusted all code accordingly and updated schema. If the user inputs a book with no decimal, it's saved to the database as a float, but is displayed as an int. If the user enters a book with a decimal it is displayed as a decimal.

+++ app/src/main/java/com/readingrecollections/d424_software_engineering_capstone/ui/entities/Book.java
@@ -31,7 +31,7 @@ public class Book implements Item {
 
     private String seriesName;
 
-    private int seriesNumber;
+    private double seriesNumber;
 
     public int getId() {
         return id;
@@ -89,11 +89,11 @@ public class Book implements Item {
         this.seriesName = seriesName;
     }
 
-    public int getSeriesNumber() {
+    public double getSeriesNumber() {
         return seriesNumber;
     }
 
-    public void setSeriesNumber(int seriesNumber) {
+    public void setSeriesNumber(double seriesNumber) {
         this.seriesNumber = seriesNumber;
     }
 }
